sample txt
