+++ sample.txt
@@ -1,0 +1,14 @@
+Inference 1 This is an old photograph of a city street with people walking on it. 
+There are buildings on either side of the street, and a horse and carriage in the foreground. 
+The people are dressed in Victorian clothing. The photograph is black and white.
+
+Inference 2 The image shows a street scene in an old city, with people walking on the sidewalk and carriages driving by.
+The buildings on either side of the street are tall and ornate, with balconies and ornate facades.
+There is a horse-drawn carriage in the foreground, and a group of people standing on the sidewalks in front of buildings. 
+The sky is overcast, and there is a sense of movement and activity in the scene.
+
+Inference 3 This is an old photograph of a city street with people walking on it.
+There are buildings on either side of the street, including a church in the background.
+The people in the photograph are dressed in Victorian Era clothing, involving top hats and long coats.
+There are also horse-drawn carriages on the street. The overall atmosphere of the photograph is one of a bustling city.
+source National Association of Librarians and Public Libraries of Romania.
